deinインストールを対話型に

+++ setup.sh
@@ -4,12 +4,14 @@ SCRIPT_DIR=$(cd $(dirname $0); pwd)
 mkdir -p ~/.config/
 ln -sfv $SCRIPT_DIR ~/.config/
 
-# ローカルファイルを作成
 mkdir -p ~/.local/share/nvim/backup
 mkdir -p ~/.local/share/nvim/undo
 
-# dein をインストール
-curl https://raw.githubusercontent.com/Shougo/dein.vim/master/bin/installer.sh > installer.sh
-sh ./installer.sh ~/.cache/dein
-rm ./installer.sh
+echo "Install dein? (y/[other])"
+read input
+if [ $input = 'y' ] ; then
+    curl https://raw.githubusercontent.com/Shougo/dein.vim/master/bin/installer.sh > installer.sh
+    sh ./installer.sh ~/.cache/dein
+    rm ./installer.sh
+fi
